% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{jackknife_influence}
\alias{jackknife_influence}
\title{Jackknife influence values}
\usage{
jackknife_influence(data, statistic)
}
\arguments{
\item{data}{A numeric vector, data frame or matrix. Resampling is over
elements (vectors) or rows (data frames and matrices).}

\item{statistic}{A function taking resampled \code{data} and returning a single
number.}
}
\value{
A numeric vector of length \code{n}.
}
\description{
Leave-one-out (jackknife) estimates of the empirical influence values of
a statistic, \eqn{L_i = (n - 1)(\bar\theta_{(\cdot)} - \hat\theta_{(i)})}.
}
\examples{
jackknife_influence(c(1, 4, 2, 8, 5), mean)
}
